add: 基底EntityのBaseEntityを作成
